++ stable 1.0.2.25

+++ AutoLedger.Domain/Entities/CarReceptionRequest.cs
@@ -1,8 +1,4 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace AutoLedger.Domain
 {
@@ -12,14 +8,12 @@ namespace AutoLedger.Domain
 
         public string Title { get; set; }
         public string Description { get; set; }
-        public long Cost { get; set; }
-
+        public decimal Cost { get; set; }
 
         public int ReceptionId { get; set; }
         public CarReception Reception { get; set; }
 
-
-        public DateTime CreatedAt { get;  set; } = DateTime.Now;
+        public DateTime CreatedAt { get; set; } = DateTime.Now;
         public DateTime UpdatedAt { get; set; } = DateTime.Now;
     }
 }
